refactor: 인터페이스 생성 및 클래스 분리
